requirements.txt
markdown
pyyaml
beautifulsoup4
jinja2